relation: traits: Generalize the `Ciruit` trait to work for all provers
