+++ Cargo.toml
@@ -13,3 +13,4 @@ repository = "https://github.com/EspressoSystems/jellyfish"
 
 [workspace.dependencies]
 itertools = { version = "0.10.1", default-features = false }
+ark-mpc = { git = "https://github.com/renegade-fi/ark-mpc" }
