+++ plonk/Cargo.toml
@@ -12,7 +12,7 @@ rust-version = { workspace = true }
 [dependencies]
 ark-ec = "0.4.0"
 ark-ff = { version = "0.4.0", features = ["asm"] }
-ark-mpc = { git = "https://github.com/renegade-fi/ark-mpc.git" }
+ark-mpc = { workspace = true }
 ark-poly = "0.4.2"
 ark-serialize = "0.4.0"
 ark-std = { version = "0.4.0", default-features = false }
