+++ relation/Cargo.toml
@@ -14,7 +14,7 @@ ark-bn254 = "0.4.0"
 ark-bw6-761 = "0.4.0"
 ark-ec = "0.4.0"
 ark-ff = { version = "0.4.0", features = ["asm"] }
-ark-mpc = { git = "https://github.com/renegade-fi/ark-mpc.git" }
+ark-mpc = { workspace = true }
 ark-poly = "0.4.0"
 ark-serialize = "0.4.0"
 ark-std = { version = "0.4.0", default-features = false }
